// src/types/multiplayer.ts
export type UserRole = 'gm' | 'player';

export interface RoomMember {
  id: string; // socket.id
  name: string;
  role: UserRole;
  color: string;
  isOnline: boolean;
}

export interface RoomPing {
  id: string;
  x: number;
  y: number;
  senderName: string;
  color: string;
  createdAt: number;
}